+++ src/database.js
@@ -1,0 +1,11 @@
+const Sequelize = require("sequelize");
+
+const sequelize = new Sequelize("database", "user", "password", {
+    host: "localhost",
+    dialect: "sqlite",
+    logging: false,
+    // SQLite only
+    storage: "database.sqlite",
+});
+
+module.exports = sequelize;
